+++ packages/react-native-editor/CHANGELOG.md
@@ -14,6 +14,7 @@ For each user feature we should also add a importance categorization label  to i
 -   [*] Remove Gallery caption button on mobile [#53010]
 -   [**] Upgrade React Native to 0.71.11 [#51303]
 -   [*] Upgrade Gradle to 8.2.1 & AGP to 8.1.0 [#52872]
+-   [*] Fix Gallery block selection when adding media [#53127]
 
 ## 1.100.1
 -   [**] Add WP hook for registering non-core blocks [#52791]